pyproject.toml
[tool.poetry]
name = "term-paper"
version = "0.1.0"
description = ""
authors = ["Anastasya Kuznetsova <kuznetsova.nastasy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pip = "^23.2.1"
requests = "^2.31.0"
datetime = "^5.2"
python-dotenv = "^1.0.0"
logging = "^0.4.9.6"
dynaconf = "^3.2.2"
fire = "^0.5.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
flake8 = "^6.1.0"
pylint = "^2.17.5"
pandas = "^2.1.0"
numpy = "^1.25.2"
mysql-connector-python = "^8.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
